refactor: extract q8 file reader scratch
